src/backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
Flask==3.1.1
flask-cors==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
python-dotenv==1.1.0
typing_extensions==4.13.2
Werkzeug==3.1.3
azure-cosmos==4.7.0
azure-identity==1.19.0 
